adjust CLI output slightly for rebuilding poller cache
Instead of spamming the terminal with dots, print a progress output
of sorts, staying on a single line. this uses bog-standard VT100
escape codes, so every terminal emulator nowadays should be
fine with that.

No check is done wether a TTY is connected on stdout,
but since this is a CLI script intended for interactive use anyway,
this should hold up (additionally, none of the CLI scripts do this,
and there's cacti_log() ).